Fix for 4 wire mode on arduino
The original #if  was meaningless becasue it would always eveluate as
true becasue LITTLEWIRE was already handled in an #elif above.  This
change allows the Tiny to still be able to use 3 wire mode, but arduinos
can use a 4 wire mode, which uses CSN but not CE (CE always held high).
Previously on an arduino if the CE and CSN pins were set to the same
pin, then CSN pinmode was never set due to this IF