Enable failure handling, add example
The three main failure modes of the radio include:
* Writing to radio: Radio unresponsive - Fixed internally by adding a
timeout to the internal write functions in RF24 (failure handling)
* Reading from radio: Available returns true always - Fixed by adding a
timeout to available functions by the user. This is implemented
internally in  RF24Network.
* Radio configuration settings are lost - Fixed by monitoring a value
that is different from the default, and re-configuring the radio if this
setting reverts to the default.