Fix ESP32 SPI speed issues
Without SPI transactions enabled, the SPI speed is set to CPU_F/128 and accessing shared devices on the SPI bus (TFT display in my case), also run at the slow speed even if transactions are used. Explicitly setting the SPI clock seems to disable the ability of transactions to affect the clock speed later. I tried fixing this in RF24_config.h, but the compiler is behaving oddly and won't properly define RF24_SPI_TRANSACTIONS, so the only way I got it to work was to place it in RF24.cpp.