[RPi] Improve exception when not running as root
It appears that on Raspberry Pi, the code checks
that the process is running as root.

When it's not, it throws a very cryptic error (see #531):
  terminate called after throwing an instance of 'int'

I'm not sure that making it compulsory to run as root
is justified, but at least this change improves the
error message.

After it appears as:
  terminate called after throwing an instance of 'std::runtime_error'
    what():  Process should run as root