refactor: migrate to `stopListening(txAddress)` from `openWritingPipe()` (#1030)
follow up to #1029

`stopListening()` now writes the TX address to pipe 0 (for reading and writing). This deprecates the need for `openWritingPipe()` by offering an overloaded `stopListening(txAddress)`.

All examples (and the python wrapper) have been updated to use the new public `stopListening(txAddress)` accordingly.

The main take away is no duplicate SPI transactions when switching to TX mode and (re)setting the TX address. This aims to fix the performance regression in #1029.

---------

Co-authored-by: TMRh20 <tmrh20@gmail.com>